move to default namespace
okteto is for dev only, and that is enforced in the admission webhooks
for ingress so that you cannot add ingresses outside of the platform
subdomain (ie. *.empfin.kingdonb.dev)

So, copy our resources to a default namespace that is not monitored by
any admission webhook enforcing ingress format structure and content!