add "default" ns - cert, ingress, issuer, secret
The DNS secret is the same as okteto-dns-secret from the okteto ns. The
cert + issuer will generate a TLS secret that is managed by cert-manager
for the kingdonb.dev and *.kingdonb.dev subject names, automatically!

Because of the way DNS is set up, these CNAMEs must be added manually.
Because of the way Commits-To works, the user rows must also be inserted
into the (Heroku-hosted, TLS-secured) Pg database manually, by an admin.

(Voila, production!)